add use prompt command doc comments
